Initial
The index.html file needs to be placed on a web server that is
accessible to the Wii U.

To configure the files, please check the comments in index.html.

The "wiiuc.jar" binary is configured to play Portal 2 (with my key
config). Change this to fit your game/keyconfig.